+++ lib/stack_log.dart
@@ -1,9 +1,9 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright (C) 2025 Uvarov Oleg <uv.ol.al@gmail.com>
 
-library google_cloud_log_parser;
-
+export 'src/config_manager.dart';
+export 'src/console_colors.dart';
 export 'src/formatters.dart';
-export 'src/filters.dart';
-export 'src/file_utils.dart';
-export 'src/color_output.dart';
+export 'src/log_highlighter.dart';
+export 'src/logger.dart';
+export 'src/output_writer.dart';