+++ CHANGELOG.md
@@ -10,3 +10,13 @@
 ## 1.0.0+1
 
 - Fixed executables names
+
+## 1.0.0+2
+
+- Fixed impossibility to write log file
+- Fixed filters not reading
+- Moved configuration to config file
+- Changed log file extension
+- Added several options to configuration
+- Translated part of output lines to English
+- Refactor